+++ src/user-action/createSessionStartedUserAction.test.ts
@@ -47,29 +47,44 @@ describe('action', () => {
     })
 
     it('sets buildId field when GRAVITY_BUILD_ID env var is set', () => {
-      process.env.GRAVITY_BUILD_ID = '51'
-      expect(createSessionStartedUserAction().buildId).toEqual('51')
+      withPatchedProperties(process.env, { GRAVITY_BUILD_ID: '51' }, () =>
+        expect(createSessionStartedUserAction().buildId).toEqual('51'),
+      )
+    })
+
+    it('sets buildId field when GRAVITY_BUILD_ID is set on the window', () => {
+      withPatchedProperties(window, { GRAVITY_BUILD_ID: '123' }, () =>
+        expect(createSessionStartedUserAction().buildId).toEqual('123'),
+      )
     })
 
     it('does not set buildId field when GRAVITY_BUILD_ID is empty', () => {
-      process.env.GRAVITY_BUILD_ID = ''
-      expect(createSessionStartedUserAction().buildId).toEqual(undefined)
+      withPatchedProperties(process.env, { GRAVITY_BUILD_ID: '' }, () =>
+        expect(createSessionStartedUserAction().buildId).toEqual(undefined),
+      )
     })
 
     it('sets buildId field when REACT_APP_GRAVITY_BUILD_ID env var is set', () => {
-      process.env.REACT_APP_GRAVITY_BUILD_ID = '42'
-      expect(createSessionStartedUserAction().buildId).toEqual('42')
+      withPatchedProperties(process.env, { REACT_APP_GRAVITY_BUILD_ID: '42' }, () =>
+        expect(createSessionStartedUserAction().buildId).toEqual('42'),
+      )
     })
 
     it('does not set buildId field when REACT_APP_GRAVITY_BUILD_ID is empty', () => {
-      process.env.REACT_APP_GRAVITY_BUILD_ID = ''
-      expect(createSessionStartedUserAction().buildId).toEqual(undefined)
+      withPatchedProperties(process.env, { REACT_APP_GRAVITY_BUILD_ID: '' }, () =>
+        expect(createSessionStartedUserAction().buildId).toEqual(undefined),
+      )
     })
 
     it('sets buildId field from GRAVITY_BUILD_ID when GRAVITY_BUILD_ID and REACT_APP_GRAVITY_BUILD_ID are set', () => {
-      process.env.GRAVITY_BUILD_ID = '12'
-      process.env.REACT_APP_GRAVITY_BUILD_ID = '13'
-      expect(createSessionStartedUserAction().buildId).toEqual('12')
+      withPatchedProperties(
+        process.env,
+        {
+          GRAVITY_BUILD_ID: '12',
+          REACT_APP_GRAVITY_BUILD_ID: '13',
+        },
+        () => expect(createSessionStartedUserAction().buildId).toEqual('12'),
+      )
     })
 
     it('returns Cypress current test if any', () => {
@@ -117,3 +132,27 @@ describe('action', () => {
     })
   })
 })
+
+function withPatchedProperties<T>(toPatch: unknown, properties: T, callBack: () => void) {
+  const originalProperties = Object.fromEntries(
+    Object.entries(properties).map(([key]) => {
+      return [key, (toPatch as T)[key]]
+    }),
+  )
+  Object.entries(properties).map(([key, value]) =>
+    Object.defineProperty(toPatch, key, { value, writable: true, configurable: true }),
+  )
+
+  try {
+    callBack()
+  } finally {
+    Object.entries(originalProperties).forEach(([key, value]) => {
+      if (value === undefined) {
+        // eslint-disable-next-line @typescript-eslint/no-dynamic-delete
+        delete (toPatch as T)[key]
+      } else {
+        Object.defineProperty(toPatch, key, { value })
+      }
+    })
+  }
+}