+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/Smartesting/gravity-data-collector/compare/v3.10.0...main)
+## [Unreleased](https://github.com/Smartesting/gravity-data-collector/compare/v3.11.0...main)
 
 ### Info
 
@@ -18,6 +18,12 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Removed
 
 ### Fixed
+
+## [3.11.0](https://github.com/Smartesting/gravity-data-collector/compare/v3.10.0...v3.11.0)
+
+### Added
+
+- Add option `buildId` to ease specifying the buildId instead of relying on global variables.
 
 ## [3.10.0](https://github.com/Smartesting/gravity-data-collector/compare/v3.9.0...v3.10.0)